git/02-switch-ssh.sh: indent with tabs; +making gitsw use dir name if no origin found

+++ git/02-switch-ssh.sh
@@ -1,14 +1,18 @@
 # Switch to SSH
 function gitsw {
-  # repo is the name of the repository
-  repo=$(git remote -v | grep origin | sed 's/.*\///g' | sed 's/.git.*//g' | sed 's/ (fetch)//g' | head -n 1)
-  git remote rm origin
+	# repo is the name of the repository
+	if git remote -v | grep origin &> /dev/null ; then
+		repo=$(git remote -v | grep origin | sed 's/.*\///g' | sed 's/.git.*//g' | sed 's/ (fetch)//g' | head -n 1)
+		git remote rm origin
+	else
+		repo=${PWD##*/}
+	fi
 
-  if [[ -n "$1" ]]; then
-      git remote add origin git@github.com:fusion809/"${1}".git
-  else
-      git remote add origin git@github.com:fusion809/"${repo}".git
-  fi
+  	if [[ -n "$1" ]]; then
+		git remote add origin git@github.com:fusion809/"${1}".git
+	else
+		git remote add origin git@github.com:fusion809/"${repo}".git
+	fi
 }
 
 alias SSH=gitsw
